[counter] `stats_counter_t` -- use `int64_t`

+++ src/packages/std/counter/stats_counter.h
@@ -10,5 +10,5 @@ void stats_counter_destroy(stats_counter_t **self_pointer);
 void stats_counter_per_thread_add1(stats_counter_t *self, size_t id);
 void stats_counter_per_thread_sub1(stats_counter_t *self, size_t id);
 
-size_t stats_counter_total(stats_counter_t *self);
+int64_t stats_counter_total(stats_counter_t *self);
 bool stats_counter_total_is_zero(stats_counter_t *self);
